ipcache: Fix node-to-node encryption with kube-apiserver
The kube-apiserver reserved identity requires us to create a IPCache
entry based on the endpoints of the `default/kubernetes` K8s service.

This kube-apiserver  IPCache entry will overwrite any existing remote
host entries. The entry also blocks the creation of remote host entries
for the same IP (via `source.AllowsOverwrite`).

When the kube-apiserver entry is created, we check if there is an
existing remote node entry, and if so, copy over the IPSec keypair
information into the newly created kube-apiserver entry. If however the
order is reversed (we first create the kube-apiserver entry before the
remote node is discovered), then the IPSec keypair is never updated.
This in turn means that connectivity to any remote nodes which also run
the kube-apiserver is broken when node-to-node encryption is enabled.

This commit works around the issue by adding a special case to the
IPCache upsert logic, allowing the remote node entry to partially
overwrite the existing kube-apiserver entry to add the required IPSec
keypair.

This work is also needed for the upcoming WireGuard node-to-node
encryption.

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>