node: Set EncryptKey if WireGuard is enabled
This commit sets the EncryptKey index field to a static non-zero value
in CiliumEndpoints and CiliumNodes if WireGuard is enabled for managing
cilium-agent. This allows us to have fine-grained tracking of encrypted
endpoints. A follow-up commit will update the datapath accordingly, to
make use of this field and only encrypt connections to endpoints which
have a non-zero encryption key index.

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>